Speed up CanonicalForm
Since "edges" is a dense list, the IsBound-check can be removed.

By experimenting I found that declaring "edge" as a variable takes some
time (which is significant as this part of the loop is executed a lot).

In total the computation time of this loop is cut in half by these
changes.